// app.client/src/api/models/CreateUserIntermediateCommand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateUserIntermediateCommand = {
    userId?: string | null;
    userEmail?: string | null;
    category?: string | null;
    categoryId?: string | null;
};